refactor(client): extract auth/whoami tests to sibling file
